コピー前に安全処理追加

+++ src/cli/init/service/copy-template.ts
@@ -5,6 +5,10 @@ export function copyTemplate(
     projectPath:string
 ):void{
 
+    if (fs.existsSync(projectPath)) {
+        throw Error(`エラー: コピー先がすでに存在するため、安全のために終了します。 (${projectPath})`);
+    }
+
     fs.cpSync(templatePath, projectPath, {
         recursive: true
     });
